requirements.txt
pydantic==1.8.2
requests==2.25.1
bs4==0.0.1
python-dateutil==2.8.1
pandas==1.3.5